// src/types/scheduler.types.ts
export interface Agent {
  id: string;
  name: string;
  avatar?: string;
  ImePrezime: string;
  isOnBreak: boolean;
  shiftStart?: Date | Timestamp | null;
  shiftEnd?: Date | Timestamp | null;
  currentBreakStartTime?: Date | Timestamp | null;
  // **FIX:** Allowed null for fields that can be cleared
  currentBreakId?: string | null;
  currentShiftId?: string | null;
  assignedBreaks?: {
    shiftId: string;
    timeSlotId: string;
    breakTypeIndex: number;
    status: 'scheduled' | 'active' | 'done';
    breakSlotId: string;
  }[];
  activeBreak_StartTime?: Date | Timestamp | null;
  activeBreak_ScheduledBreakId?: string | null;
  totalBreakDurationToday?: number | null;
}